Smartcard API v3.0: Unregister all the receivers in the onDestroy clbk
BZ: 135475

All the broadcast receivers which are registered in the onPostExcecute
callback of the  initialization async task should be properly
unregistered in the onDestroy callback of the SmartcardService.

The onPostExecute callback is called after the onCreate callback.
The onDestroy callback can be called at any time after the onCreate
callback (so potentially BEFORE the completion of the initialisation async task).
That's why the initialization async task should also be canceled in the
onDestroy clbk in order to prevent onPostExecute from being called later.

Category: feature differentiation
Domain: CWS.NFC-SEEK
Origin: internal
Upstream-Candidate: no, proprietary

Change-Id: I5c7d2d4f94903cb41f597ce92aaf388d41619d5d
Signed-off-by: Sebastien Griffoul <sebastien.griffoul@intel.com>
Reviewed-on: http://android.intel.com:8080/132479
Reviewed-by: Niciarz, Sebastian <sebastian.niciarz@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>