Smartcard API v3.0: Fallback to ARF each time the ARA selection fails
BZ: 139549

The GP spec says that the ARF should be used ONLY if the ARA-M applet is not found
(error 6A). Access to the SIM card should be denied in all the other cases.

However we found few SIM cards (from Oberthur) whithout any ARA-M applet and which
return the error code 6999 instead of 6A82 when an unknown applet is being selected.

In order to support this kind of card we need to be less strict than the GP spec:
If any error occurs during ARA-M selection then the ACE should act as if the ARA-M is not
found.

Category: feature differentiation
Domain: CWS.NFC-SEEK
Origin: internal
Upstream-Candidate: no, proprietary

Change-Id: If1cb1582e39fc611c9681be9083bee21cc44ee51
Signed-off-by: Sebastien Griffoul <sebastien.griffoul@intel.com>
Reviewed-on: http://android.intel.com:8080/132950
Reviewed-by: Niciarz, Sebastian <sebastian.niciarz@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>