Smartcard API v3.1: ACCF parser is not fully compliant with GP spec
BZ: 148983

The ACCF parser deosn't support the case where:
- there is no application hash.
- but there is an APDU  rules.

Moreover this parser is not lenient and is not able to understand
the B-TLV sequence 30 0E:
                     A0 0C:
                        A0 05:
                          A0 03:  <-- object not handled by the parser
                            80 01: 00
                         A1 03:
                            80 01: 01

This is a regression introduced by the patch made for issue 64
(https://code.google.com/p/seek-for-android/issues/detail?id=64).
In order to be able to correctly parse any kind of ACCF rules
the parser should be lenient and handle both cases.

Category: feature differentiation
Domain: CWS.NFC-SEEK
Origin: internal
Upstream-Candidate: No, not-aosp

Change-Id: Ia8a1f59e0bfe5c08ef1cdaca1fb64a5fdb55fcb5
Signed-off-by: Sebastien Griffoul <sebastien.griffoul@intel.com>